feat: updated export
