+++ src/index.ts
@@ -1,26 +1,6 @@
-import { addBookmark } from './formatters/bookmark';
-import { addFolder } from './formatters/folder';
-import { indentLeft, isBookmark, isFolder } from './utils';
+import { buildHtml } from './buildHtml';
 import { HEADER, DEFAULT_AMOUNT_TO_PAD_BY } from './constants';
 import { Bookmark, Folder } from './ts/types';
-
-export const buildHtml = (
-  content: (Bookmark | Folder)[],
-  indent: number,
-): string =>
-  content
-    .reduce((html: string[], current) => {
-      if (isFolder(current)) {
-        html.push(addFolder(current, indentLeft(indent)));
-      }
-
-      if (isBookmark(current)) {
-        html.push(addBookmark(current, indentLeft(indent)));
-      }
-
-      return html;
-    }, [])
-    .join('\n');
 
 export const bookmarked = (content?: (Bookmark | Folder)[]): string =>
   [
