+++ tsconfig.json
@@ -20,5 +20,5 @@
     ]
   },
   "include": ["src/**/*"],
-  "exclude": ["node_modules", "*.test.*"]
+  "exclude": ["node_modules", "*.test.ts", "src/test/"]
 }
